# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: >
    A multi-tenant SaaS backend where tenant isolation is structural,
    enforced by PostgreSQL Row-Level Security and composite foreign keys.
  version: 1.0.0
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      operationId: signup
      tags: [Auth]
      summary: Atomically create an organisation, its first ORG_ADMIN user, and issue a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Organisation and admin user created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: Organisation slug or user email already taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Exchange email/password for a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials or inactive account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      summary: Resolve the bearer token to the current user and organisation
      description: >
        Re-reads the caller's row from the database rather than trusting the
        token's claims: a JWT keeps asserting the role and account it was
        minted with until it expires, so a user deactivated or demoted in the
        meantime must not keep their old access. Call this on app boot to
        restore a session — never decode the JWT in the browser to decide
        what a user may do.
      responses:
        '200':
          description: The authenticated user and their organisation
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  organisation: { $ref: '#/components/schemas/Organisation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Record a logout
      description: >
        Access tokens are stateless, so there is nothing to revoke
        server-side — logging out is the client discarding its token. This
        endpoint exists so the client has one obvious call to make and the
        event reaches the audit log.
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List the caller's own organisation's members
      description: >
        Readable by any authenticated member, not only admins — members need
        the directory to see colleagues and pick assignees. Managing those
        users stays admin-only. Always scoped to the caller's organisation.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Organisation members
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createUser
      tags: [Auth]
      summary: Add a member to the caller's own organisation
      description: >
        Requires ORG_ADMIN (or PLATFORM_ADMIN). The organisation is always
        the caller's own, taken from their JWT — it is never accepted as
        request input. Enforces the plan's seat limit atomically via
        UsageCounter; returns 409 when the org is already at max_users.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Email already registered, or seat limit reached
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /users/{id}:
    get:
      operationId: getUser
      tags: [Users]
      summary: One member of the caller's own organisation
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateUser
      tags: [Users]
      summary: Update a member's name, role, or status
      description: >
        ORG_ADMIN only. A user in another organisation is invisible to RLS
        and so returns 404, never 403 — the response must not confirm that
        the id exists. PLATFORM_ADMIN can never be assigned here.
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: The updated member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Would leave the organisation with no active administrator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Remove a member and free their seat
      description: >
        ORG_ADMIN only. Releases the USERS quota the account was holding.
        Refused if the caller targets themselves, or if it would leave the
        organisation with no active administrator.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: Removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Self-deletion, or would leave no active administrator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /organisations/{orgId}/users:
    get:
      operationId: listOrganisationUsers
      tags: [Platform Admin]
      summary: List one organisation's members
      description: >
        Served over a read-only DB connection (the `platform_reader` role)
        with SELECT grants limited to Organisation/SubscriptionPlan/User —
        it cannot see Project, Issue, or UsageCounter rows even if asked to.
      parameters:
        - $ref: '#/components/parameters/OrganisationIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Members of the organisation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Obtained from POST /api/auth/signup or POST /api/auth/login. Payload is { userId, orgId, role }.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      required: false
      description: Rows per page. Capped so one request cannot pull a whole table into memory.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    OrganisationIdPath:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string, example: VALIDATION_ERROR }
              fields: { type: object, additionalProperties: true }
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but the caller's role cannot perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >
        Not found — either the resource genuinely doesn't exist, or it
        belongs to a different organisation than the caller's (Row-Level
        Security filters those rows out before the app ever sees them, so
        both cases are indistinguishable from outside, by design).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    PaginationEnvelope:
      type: object
      description: >
        The shape every list endpoint returns. Rows are filtered and paged in
        PostgreSQL — the API never loads a full set into memory to slice it.
      required: [items, pagination]
      properties:
        items:
          type: array
          items: {}
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page: { type: integer, example: 1 }
            limit: { type: integer, example: 20 }
            total: { type: integer, description: Total matching rows across all pages, example: 120 }
            totalPages: { type: integer, example: 6 }
    UpdateUserRequest:
      type: object
      minProperties: 1
      description: At least one field is required. PLATFORM_ADMIN is intentionally not assignable.
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        role: { type: string, enum: [ORG_ADMIN, ORG_MEMBER] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
...
    Error:
      type: object
      properties:
        code: { type: string, example: CONFLICT }
        message: { type: string }
    SignupRequest:
      type: object
      required: [organisationName, organisationSlug, email, password, firstName, lastName]
      properties:
        organisationName: { type: string, minLength: 2, maxLength: 120 }
        organisationSlug: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        organisation: { $ref: '#/components/schemas/Organisation' }
        user: { $ref: '#/components/schemas/User' }
    CreateUserRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string, enum: [ORG_ADMIN, ORG_MEMBER], default: ORG_MEMBER }
    Organisation:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [ACTIVE, SUSPENDED] }
        plan_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        organisation_id: { type: string, format: uuid }
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        role: { type: string, enum: [PLATFORM_ADMIN, ORG_ADMIN, ORG_MEMBER] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...